+++ Documents/NetBeansProjects/Prueba2ICLProgra2/CHANGELOG.md
@@ -1,5 +1,6 @@
 1.0: Codigo base: Entidades vehiculo y crearfile incompleto.
-1.1: Codigo base: CrearFile Corregido. Presentacion incompleta.
-
+1.1: Codigo base: CrearFile Corregido. Presentacion incompleta. Logica creada con los metodos para el manejo del CRUD.
+1.2: Se agrega la capa presentacion completa, con las pestañas para el usuario final, se corrige la presentacion
+de los nanosegundos.